fix: tsc-esm and moving files
