+++ packages/core/base/src/SolanaNextAuth.ts
