--- packages/core/base/src/adapters/index.ts
@@ -1,1 +1,0 @@
-export * from "./firebase";
