--- packages/core/base/src/handlers/firebase/index.ts
@@ -1,1 +1,0 @@
-export * from "./SolanaFirebaseAuth";
