--- packages/core/base/src/handlers/index.ts
@@ -1,3 +1,0 @@
-export * from "./firebase";
-export * from "./SolanaAuth";
-export * from "./SolanaNextAuth";
