+++ packages/core/base/src/index.ts
@@ -1,3 +1,5 @@
-export * from "./handlers";
-export * from "./adapters";
+export * from "./firebase";
+export * from "./SolanaNextAuth";
+export * from "./SolanaFirebaseAuth";
+export * from "./SolanaAuth";
 export * from "./config";
