+++ packages/ui/react-ui/src/SolanaAuthButtonFirebase.tsx
